Move ParseImage to backend
